Make the `scribble/html' layout mimic the one for `scribble/text'.
Ie, the difference between using it as a language and as a module is now
the same in both.  Also, improve `scribble/html' (and
`scribble/html/htmllang') by reproviding most of the corresponding text
modules.

Change `meta/web' accordingly, and improve code by making a new langugae
(`#lang meta/web').  This language is similar to `scribble/html' except
that it uses the plain scribble reader (not the `-inside' one), it
doesn't use the customized module-begin feature (that uses
`output-xml'), and it adds all the relevant bits of `meta/web/common'.
(Also, "meta/web/common.rkt" is gone, since it's used only as a language
now.)

This commit has lots of details and binding games, so it's tricky, and
potentially caused some problems.  (Took me a while to track many
breakages, so I won't be surprised if there are more.)

original commit: 82221a51db9648d74caef31837b1f00b4e9d5773