More head-level additions.
Note that I used "image/x-icon" for the "shortcut icon" type, even
though it's different from the one used on the "icon" type right above
it.

Need to sort the CSS mess: I think that "css/style.css" is meant to be
out global CSS (so it needs to change), and "css/scribble.css" is
probably what "more.css" is doing and it's being used only on the front
page since that has the scribble-rendered examples.

Also, need to scan all the "@;" comments here: some might need to be in
the actual HTML output, and some look like leftovers from wherever this
HTML was ripped from that are not applicable.

original commit: 5ed2101d50c0f7a2469c1b9ad3fe021a9c251a56