main doc page: use `margin-note` instead of `margin-note*`
The rendered form of `margin-note` does the right thing with a narrow
view, unlike the new rendering of `margin-note*`.

Also, wrap `not-on-the-web` only around the part that claims
an installation-specific listing, and add `not-on-the-web` to
"doc-site.css" for "docs.racket-lang.org". I think it's good
for :docs.racket-lang.org" to explain how to open documentation
locally, which is why I didn't make `not-on-the-web` work
as a block wrapper.

original commit: e9b64fce2080605337a03d0aecf247e4a5cc6157