A basic plan takes shape
Use @media to condition on the width -- below a certain threshold we
enter a vertical-scrolling mode with all of the content, so the menus
and clicking are gone. (This seems to be the way that the Julia
website works.)

Lots of work still needed to tune the page so that it looks right in
both modes but this approach doesn't seem impossible to make work.