download: fix <select> not reset on Firefox
On Firefox, setAttribute('selected') can't be used to programmatically
update the page. Apparently, this appears to be the correct behavior
according to the HTML specification, and other browsers (e.g., Chrome)
do it wrong.
(See https://github.com/jsdom/jsdom/issues/1130#issuecomment-102375898)
In any case, the solution here is to mutate the HTML object directly
(that is, change the property, not the attribute).