Style Racket code blocks used on the blog
This copies in the essential styles from Scribble to present Racket code blocks
with nice formatting on the blog. In particular, this ensures they are shown
with monospaced fonts and use the font colours from the documentation.

The styles added here are deliberately kept minimal to avoid changing the blog's
overall look and feel outside of Racket code blocks.

Fixes https://github.com/racket/racket-lang-org/issues/228