Add "version-redirect.rkt" script for pruning global redirects
It's convenient for a release to just redirect all
`releases/<vers>/installers` requests at `download.racket-lang.org` to
`mirror.racket-lang.org`, but we can only link up to 50 versions that
way through a bucket redirect rule. For old versions, make an
indivudual bucket object for each redirection that should work (and
there are a small number, because it's roughlyone per installer).

For now, versions before v7.0 have been changed in this way, which
buys us another 4 years or so before we run out of redirections again.
Assuming no problems are desovered in the next year or so, we can turn
it up to before v9.0, which will put off the problem for a decae or
so. And we can keep turning up the version this way indefinitely in
the future.